Refactor: Moved all `View` styling methods.
