+++ Client/ts/Components/TextField.ts
@@ -6,4 +6,19 @@ export default class TextField extends InputField {
         super(placeholder || '');
         this.body.type = 'text';
     }
+
+    textStart(): this {
+        this.body.style.textAlign = 'left';
+        return this;
+    }
+
+    textCenter(): this {
+        this.body.style.textAlign = 'center';
+        return this;
+    }
+
+    textEnd(): this {
+        this.body.style.textAlign = 'right';
+        return this;
+    }
 }
